Move `utils update-cluster-stack` to `update cluster`
