+++ pkg/ctl/update/update.go
@@ -1,0 +1,24 @@
+package update
+
+import (
+	"github.com/kris-nova/logger"
+	"github.com/spf13/cobra"
+	"github.com/weaveworks/eksctl/pkg/ctl/cmdutils"
+)
+
+// Command will create the `create` commands
+func Command(g *cmdutils.Grouping) *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "update",
+		Short: "Update resource(s)",
+		Run: func(c *cobra.Command, _ []string) {
+			if err := c.Help(); err != nil {
+				logger.Debug("ignoring error %q", err.Error())
+			}
+		},
+	}
+
+	cmd.AddCommand(updateClusterCmd(g))
+
+	return cmd
+}
